Rearranging views from instagram_clone
